refactor, create vyper/context/bases subpackage
