+++ vyper/context/types/bases/structure.py
